+++ bundle/manifests/ibm-licensing-operator.clusterserviceversion.yaml
@@ -62,7 +62,7 @@ metadata:
     categories: Monitoring
     certified: "false"
     containerImage: icr.io/cpopen/ibm-licensing-operator:4.0.0
-    createdAt: "2023-04-12T11:49:12Z"
+    createdAt: "2023-04-12T12:33:33Z"
     description: The IBM Licensing Operator provides a Kubernetes CRD-Based API to monitor the license usage of products.
     olm.skipRange: '>=1.0.0 <4.0.0'
     operatorframework.io/suggested-namespace: ibm-licensing
@@ -592,16 +592,6 @@ spec:
                 - update
                 - watch
             - apiGroups:
-                - operator.com.ibm
-              resources:
-                - namespacescope
-                - namespacescope/finalizers
-                - namespacescope/status
-              verbs:
-                - get
-                - list
-                - watch
-            - apiGroups:
                 - operator.ibm.com
               resources:
                 - ibmlicensings
@@ -614,6 +604,16 @@ spec:
                 - list
                 - patch
                 - update
+                - watch
+            - apiGroups:
+                - operator.ibm.com
+              resources:
+                - namespacescope
+                - namespacescope/finalizers
+                - namespacescope/status
+              verbs:
+                - get
+                - list
                 - watch
             - apiGroups:
                 - operators.coreos.com
