+++ config/rbac/role.yaml
@@ -163,16 +163,6 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operator.com.ibm
-  resources:
-  - namespacescope
-  - namespacescope/finalizers
-  - namespacescope/status
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
   - operator.ibm.com
   resources:
   - ibmlicensings
@@ -185,6 +175,16 @@ rules:
   - list
   - patch
   - update
+  - watch
+- apiGroups:
+  - operator.ibm.com
+  resources:
+  - namespacescope
+  - namespacescope/finalizers
+  - namespacescope/status
+  verbs:
+  - get
+  - list
   - watch
 - apiGroups:
   - operators.coreos.com
